backend/requirements.txt
flask
mysql-connector-python
sqlalchemy
pandas
